+++ virtualizarr/tests/test_parsers/test_hdf/test_hdf_integration.py
@@ -9,6 +9,7 @@ from virtualizarr.tests import (
     requires_icechunk,
     requires_imagecodecs,
     requires_kerchunk,
+    requires_netcdf4,
 )
 from virtualizarr.tests.test_integration import roundtrip_as_in_memory_icechunk
 
@@ -16,6 +17,7 @@ from virtualizarr.tests.test_integration import roundtrip_as_in_memory_icechunk
 @requires_kerchunk
 @requires_hdf5plugin
 @requires_imagecodecs
+@requires_netcdf4
 class TestIntegration:
     def test_filters_h5netcdf_roundtrip(
         self, tmp_path, filter_encoded_roundtrip_hdf5_file, local_registry